#6 added onClick event handler to show weather data when user clicks 'weather' link, created function component to create stucture for weather modal, edited db to save weather data as json and not text
