+++ client/components/home.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { CreateWeatherModal } from './create-weather-modal';
 
 export default class Home extends React.Component {
   constructor(props) {
@@ -47,7 +48,6 @@ export default class Home extends React.Component {
     this.handleClickCloseWeatherModal = this.handleClickCloseWeatherModal.bind(this);
     this.getLocalDateString = this.getLocalDateString.bind(this);
     this.getIsoDates = this.getIsoDates.bind(this);
-    this.modalInfo = this.modalInfo.bind(this);
     this.handleChangeSurfBox = this.handleChangeSurfBox.bind(this);
     this.handleClickEntriesModal = this.handleClickEntriesModal.bind(this);
     this.handleChangeEntriesTextbox = this.handleChangeEntriesTextbox.bind(this);
@@ -194,49 +194,6 @@ export default class Home extends React.Component {
     }
   }
 
-  modalInfo(data) {
-    if (data) {
-      const sunrise = data.daily.sunrise.toString();
-      const sunset = data.daily.sunset.toString();
-      const rain = data.daily.precipitation_sum.toString();
-      const maxTemp = data.daily.temperature_2m_max.toString();
-      const minTemp = data.daily.temperature_2m_min.toString();
-      const windSpeed = data.daily.windspeed_10m_max.toString();
-      const windGusts = data.daily.windgusts_10m_max.toString();
-      const modalWindowInfo = (
-    <>
-      <div className='row modal-info-row'>
-        <div className='col-6 left-col'>
-          <p><i className="fa-regular fa-sun"></i> sunrise time:</p>
-          <p><i className="fa-regular fa-sun"></i> sunset time:</p>
-          <br></br>
-          <p><i className="fa-solid fa-cloud-showers-heavy"></i> rain amount:</p>
-          <br></br>
-          <p><i className="fa-solid fa-temperature-arrow-up"></i> max. temp.:</p>
-          <p><i className="fa-solid fa-temperature-arrow-down"></i> min. temp.:</p>
-          <br></br>
-          <p><i className="fa-solid fa-wind"></i> wind speed:</p>
-          <p><i className="fa-solid fa-wind"></i> wind gusts:</p>
-        </div>
-        <div className='col-6 right-col'>
-          <p>{`${sunrise.substring(12)} AM`}</p>
-          <p>{`${sunset.substring(12)} PM`}</p>
-          <br></br>
-          <p>{`${rain} inches`}</p>
-          <br></br>
-          <p>{`${maxTemp} °F`}</p>
-          <p>{`${minTemp} °F`}</p>
-          <br></br>
-          <p>{`${windSpeed} mph`}</p>
-          <p>{`${windGusts} mph`}</p>
-        </div>
-      </div>
-    </>
-      );
-      return modalWindowInfo;
-    }
-  }
-
   handleChangeSurfBox(isChecked, index) {
     const copyDaysArr = this.state.days.slice();
     const copyDay = Object.assign({}, copyDaysArr[index]);
@@ -308,7 +265,7 @@ export default class Home extends React.Component {
           <div className='modal-window'>
             <button className='modal-button' onClick={this.handleClickCloseWeatherModal}><i className="fa-solid fa-xmark"></i></button>
             <h4 className='mt-2 mb-4'>Weather Forecast</h4>
-            {this.modalInfo(this.state.entryValueWeather)}
+            {CreateWeatherModal(this.state.entryValueWeather)}
           </div>
         </div>
         <div className={entriesModalClass}>
