+++ client/components/past-logs.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { CreateWeatherModal } from './create-weather-modal';
 
 const styles = {
   row: {
@@ -10,8 +11,12 @@ export default class PastLogs extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
-      entries: null
+      entries: null,
+      isClickedWeather: false,
+      targetDayWeather: null
     };
+
+    this.handleClickWeatherModal = this.handleClickWeatherModal.bind(this);
   }
 
   componentDidMount() {
@@ -32,7 +37,7 @@ export default class PastLogs extends React.Component {
               <div className='text-start'>
                 <p className='entry-content'>{day.content}</p>
               </div>
-              <a className='weather-link mt-5' id={index}>
+              <a className='weather-link mt-5' id={index} onClick={this.handleClickWeatherModal}>
                 <i className="fa-solid fa-cloud mt-5"></i>
                 The Weather
                 <i className="fa-solid fa-cloud mt-5"></i>
@@ -44,15 +49,43 @@ export default class PastLogs extends React.Component {
     });
   }
 
+  handleClickWeatherModal(event) {
+    const targetDay = this.state.entries[event.target.id];
+    if (this.state.isClickedWeather) {
+      this.setState({
+        isClickedWeather: false,
+        targetDayWeather: null
+      });
+    } else {
+      this.setState({
+        isClickedWeather: true,
+        targetDayWeather: targetDay.weather
+      });
+    }
+  }
+
   render() {
     if (this.state.entries === null) {
       return null;
+    }
+    let modalBackgroundClass;
+    if (this.state.isClickedWeather) {
+      modalBackgroundClass = 'modal-background';
+    } else {
+      modalBackgroundClass = 'modal-background hidden';
     }
     return (
       <>
         <div className='container'>
           <h2 className='mb-5 mt-3'>Past Logs</h2>
           <div style={styles.row} className='row row-cols-md-2 row-cols-lg-4'>{this.createPastLogCards()}</div>
+        </div>
+        <div className={modalBackgroundClass}>
+          <div className='modal-window'>
+            <button className='modal-button' onClick={this.handleClickWeatherModal}><i className="fa-solid fa-xmark"></i></button>
+            <h4 className='mt-2 mb-4'>Weather Forecast</h4>
+            {CreateWeatherModal(this.state.currentDayWeather)}
+          </div>
         </div>
       </>
     );
